Tests are done
Final version is ready

+++ src/main/java/com/workintech/s18d1/dao/BurgerDaoImpl.java
@@ -2,12 +2,14 @@ package com.workintech.s18d1.dao;
 
 import com.workintech.s18d1.entity.BreadType;
 import com.workintech.s18d1.entity.Burger;
+import com.workintech.s18d1.exceptions.BurgerException;
+import com.workintech.s18d1.util.BurgerValidation;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
-import jakarta.persistence.Query;
 import jakarta.persistence.TypedQuery;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -15,26 +17,40 @@ import java.util.List;
 @Repository
 public class BurgerDaoImpl implements BurgerDao{
 
-    @PersistenceContext
-    private EntityManager entityManager;
+    private final EntityManager entityManager;
+
+    private final BurgerValidation burgerValidation;
 
 
-//    @Autowired
-//    public BurgerDaoImpl(EntityManager entityManager)
-//    {
-//        this.entityManager = entityManager;
-//    }
+    @Autowired
+    public BurgerDaoImpl(BurgerValidation burgerValidation, EntityManager entityManager)
+    {
+        this.burgerValidation = burgerValidation;
+        this.entityManager = entityManager;
+    }
 
     @Transactional
     @Override
     public Burger save(Burger burger) {  // Now I have to add some exception handler mechanism
+        burgerValidation.validateBurgerForSave(burger);
+//        if(entityManager.find(Burger.class, burger.getId()) != null)
+//        {
+//            throw new BurgerException("This id belongs to another burger", HttpStatus.BAD_REQUEST);
+//        }
         entityManager.persist(burger);
         return burger;
     }
 
     @Override
     public Burger findById(Long id) {
-        return entityManager.find(Burger.class, id);
+        burgerValidation.validateId(id);
+        Burger burgerFound = entityManager.find(Burger.class, id);
+
+        if(burgerFound == null)
+        {
+            throw new BurgerException("Burger not found", HttpStatus.NOT_FOUND);
+        }
+        return burgerFound;
     }
 
     @Override
@@ -44,37 +60,58 @@ public class BurgerDaoImpl implements BurgerDao{
     }
 
     @Override
-    public List<Burger> findByPrice(Double price) {
-
-        String sql = "SELECT * FROM fsweb_18_4.findByPrice(:price)";
-
+    public List<Burger> findByPrice(int price) {
+        double dPrice = price;
+        burgerValidation.validateFindByPrice(dPrice);
         return entityManager
-                .createNativeQuery(sql, Burger.class)
-                .setParameter("price", price)
+                .createQuery(
+                        "SELECT b FROM Burger b WHERE b.price > :price ORDER BY b.price DESC",
+                        Burger.class
+                )
+                .setParameter("price", dPrice)
                 .getResultList();
     }
 
 
+
     @Override
     public List<Burger> findByBreadType(BreadType breadType) {
-        return entityManager.createNativeQuery("SELECT * FROM fsweb_18_4.findByBreadType(:breadType)", Burger.class).setParameter("breadType", breadType.name()).getResultList();
+        burgerValidation.validateBreadType(breadType);
+
+        TypedQuery<Burger> query = entityManager.createQuery(
+                "SELECT b FROM Burger b WHERE b.breadType = :breadType ORDER BY b.name ASC",
+                Burger.class
+        );
+        query.setParameter("breadType", breadType);
+        return query.getResultList();
     }
 
     @Override
     public List<Burger> findByContent(String contents) {
-        return List.of();
+        burgerValidation.validateContent(contents);
+        TypedQuery<Burger> query = entityManager.createQuery("SELECT b FROM Burger b WHERE b.content = :contents",
+                Burger.class);
+        query.setParameter("contents", contents);
+        return query.getResultList();
     }
 
     @Transactional
     @Override
     public Burger update(Burger burger) {
+        burgerValidation.validateBurgerForUpdate(burger.getId(), burger);
+        Burger burgerFound = entityManager.find(Burger.class, burger.getId());
         return entityManager.merge(burger);
     }
 
     @Transactional
     @Override
     public Burger remove(Long id) {
+        burgerValidation.validateId(id);
         Burger burger = entityManager.find(Burger.class, id);
+        if(burger == null)
+        {
+            throw new BurgerException("Burger is not found with id: "+ id, HttpStatus.NOT_FOUND);
+        }
         entityManager.remove(burger);
         return burger;
     }
